🚧  Initial solution for adding full-height class
The modal-wrapper component and its ion-modal parent
is now separated by a shadow-dom, which means that nativeElement.closest
no longer works, and an alternative approach for adding the full-height
css class is needed.

The solution has also been slightly refactored so the class is added to
the ion-modal element instead, now the using custom css properties of
that component.